src/multi_agent/mcp_sse_server/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-projects>=1.0.0b11
azure-identity
azure-storage-blob>=12.19.0
python-pptx>=0.6.21
requests>=2.31.0
Pillow
python-dotenv
reportlab>=4.0.0
PyMuPDF>=1.23.0

